feat: Refactored wireguard so that hosts are globally accessible
Comment by Quarry:
error: The option `networking.ownWireguard.hosts.wireguardController.options' does not exist. Definition values:
- In `/nix/store/<source-hash>-source/modules/wireguard/default.nix':
           {
             IP = {
               _type = "option";
               description = "IPv4 of the wireguard client";
               type = {
           ...
was caused by
WireguardHost {
      mainIP = (generateLastIPDigit mainPrefix lastIPDigit);
      main = (generateWireguardPeer mainPrefix lastIPDigit mainPublicKey);
      unlock = (generateWireguardPeer unlockPrefix lastIPDigit unlockPublicKey);
};
instaed of
{
      mainIP = (generateLastIPDigit mainPrefix lastIPDigit);
      main = (generateWireguardPeer mainPrefix lastIPDigit mainPublicKey);
      unlock = (generateWireguardPeer unlockPrefix lastIPDigit unlockPublicKey);
};

This way we attempted to construct a submodule with a set(we think) that
did not contain the options type, resulting in the afformentioned error,
as the resulting attrset did not have an options value, put only e.g.
the main or IP, depending on WireguardPeer or WireguardHost.
Removing this uses the normal constructor of the submodule, which does
nix magic under the hood and wires it up correctly. May attempt to set
it manually with this later:
{
  options = {
      mainIP = (generateLastIPDigit mainPrefix lastIPDigit);
      main = (generateWireguardPeer mainPrefix lastIPDigit mainPublicKey);
      unlock = (generateWireguardPeer unlockPrefix lastIPDigit unlockPublicKey);
  };
};